file dates and reorg
