+++ .gitignore
@@ -1,6 +1,7 @@
 resources/webpages/freddiewoodruff.co.uk/carina.png
 resources/webpages/freddiewoodruff.co.uk/cv.pdf
 resources/webpages/freddiewoodruff.co.uk/london.mp4
+resources/webpages/freddiewoodruff.co.uk/.well-known
 target/
 objects/
 .vscode
