+++ src/Application/http_handler.hpp
@@ -17,7 +17,7 @@ namespace fbw {
 
 [[nodiscard]] task<bool> application_handler(http_ctx& connection);
 [[nodiscard]] task<bool> redirect_handler(http_ctx& connection);
-task<void> send_error(http_ctx& connection, uint32_t status_code, std::string status_message);
+
 
 } // namespace fbw
