+++ src/HTTP/HTTP2/h2frame.cpp
@@ -1,3 +1,9 @@
+//
+//  h2frame.cpp
+//  HTTPS Server
+//
+//  Created by Frederick Benjamin Woodruff on 12/10/2024.
+//
 
 #include "h2frame.hpp"
 #include "../../global.hpp"
