+++ src/HTTP/HTTP2/h2frame.hpp
@@ -1,8 +1,8 @@
 //
-//  HTTP2.hpp
+//  h2frame.cpp
 //  HTTPS Server
 //
-//  Created by Frederick Benjamin Woodruff on 26/07/2024.
+//  Created by Frederick Benjamin Woodruff on 12/10/2024.
 //
 
 #ifndef http2frame_hpp
