+++ src/HTTP/HTTP2/hpack.cpp
@@ -9,7 +9,7 @@
 #include "h2frame.hpp"
 #include "h2proto.hpp"
 
-#include "../HTTP1_1/string_utils.hpp"
+#include "../string_utils.hpp"
 
 namespace fbw {
