+++ src/Runtime/concurrent_queue.cpp
@@ -1,10 +1,11 @@
-
-#ifndef concurrent_queue_hpp
-#define concurrent_queue_hpp
+//
+//  concurrent_queue.hpp
+//  HTTPS Server
+//
+//  Created by Frederick Benjamin Woodruff on 7/7/2024.
+//
 
 #include "concurrent_queue.hpp"
 
 
 char sentinel_byte {};
-
-#endif // concurrent_queue_hpp
